requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt --no-annotate --python-version 3.12 --python-platform aarch64-apple-darwin
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
bitarray==3.8.1
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
ckzg==2.1.7
click==8.3.3
cytoolz==1.1.0
datasets==4.8.4
dill==0.4.1
eth-abi==5.2.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
fastapi==0.136.1
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
gitpython==3.1.47
gql==4.0.0
graphql-core==3.2.8
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.12.0
hyperframe==6.1.0
idna==3.13
iniconfig==2.3.0
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lomond==0.3.3
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mlx==0.29.3
mlx-lm==0.29.1
mlx-metal==0.29.3
mlx-vlm==0.3.9
multidict==6.7.1
multiprocess==0.70.19
narwhals==2.20.0
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.2
pandas==3.0.2
parsimonious==0.10.0
pillow==12.2.0
pluggy==1.6.0
poly-eip712-structs==0.0.1
polymarket-apis==0.5.8
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
py-order-utils==0.3.2
pyarrow==24.0.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.13.3
pydantic-core==2.46.3
pydeck==0.9.2
pygments==2.20.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyunormalize==17.0.0
pyyaml==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rlp==4.1.0
rpds-py==0.30.0
safetensors==0.7.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soundfile==0.13.1
soupsieve==2.8.3
starlette==1.0.0
streamlit==1.56.0
tenacity==9.1.4
tokenizers==0.22.2
toml==0.10.2
toolz==1.1.0
tornado==6.5.5
tqdm==4.67.3
transformers==5.6.2
typer==0.25.0
types-requests==2.33.0.20260408
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.46.0
web3==7.15.0
websockets==15.0.1
xxhash==3.7.0
yarl==1.23.0
